feat(server): implement core offline-first synchronization engine
- Setup Ktor backend with Koin (DI) and Exposed (ORM).
- Configure PostgreSQL with Docker Compose and Flyway migrations.
- Implement Clean Architecture layers (API, Core, Data).
- Add 'Push' flow: supports upsert and soft-delete for budgets and outgoings.
- Add 'Pull' flow: implements incremental delta sync using server-side
  timestamps and 'since' cursor logic.
- Centralize time management with Kotlin extension mappers for
  seamless conversion between Java OffsetDateTime and Kotlin Instant.
- Finalize multiplatform network DTO integration for both flows.

+++ server/data/src/main/resources/db/migration/V1__initial_schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE users (
 CREATE TABLE budgets (
     id VARCHAR(36) PRIMARY KEY, -- L'ID généré par le mobile
     user_id VARCHAR(128) NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-    monthly_income_in_cents INTEGER NOT NULL DEFAULT 0,
+    monthly_income_in_cents BIGINT NOT NULL DEFAULT 0,
     server_updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
 
@@ -18,7 +18,7 @@ CREATE TABLE outgoings (
     budget_id VARCHAR(36) NOT NULL REFERENCES budgets(id) ON DELETE CASCADE,
     user_id VARCHAR(128) NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     name VARCHAR(255) NOT NULL,
-    amount_in_cents INTEGER NOT NULL,
+    amount_in_cents BIGINT NOT NULL,
     recurrence VARCHAR(50) NOT NULL,
     due_day INTEGER NOT NULL,
     due_month INTEGER, -- Optionnel (null pour mensuel)
